+++ package.json
@@ -70,7 +70,7 @@
     "babel-eslint": "^6.0.4",
     "babel-plugin-import": "^1.0.0",
     "babel-plugin-transform-runtime": "~6.15.0",
-    "bisheng": "^0.12.0",
+    "bisheng": "^0.15.0",
     "bisheng-plugin-antd": "~0.2.0",
     "bisheng-plugin-description": "^0.1.1",
     "bisheng-plugin-react": "^0.1.0",
@@ -132,10 +132,10 @@
     "build": "npm run dist",
     "compile": "antd-tools run compile",
     "clean": "antd-tools run clean",
-    "start": "bisheng start -c ./site/bisheng.config.js",
-    "site": "bisheng build -c ./site/bisheng.config.js",
-    "just-deploy": "npm run clean && bisheng gh-pages -c ./site/bisheng.config.js",
-    "deploy": "npm run clean && bisheng gh-pages -c ./site/bisheng.config.js",
+    "start": "concurrently \"bisheng start -c ./site/bisheng.site.config.js --no-livereload\" \"bisheng start -c ./site/bisheng.kitchen.config.js --no-livereload\"",
+    "site": "concurrently \"bisheng build -c ./site/bisheng.site.config.js\" \"bisheng build -c ./site/bisheng.kitchen.config.js\"",
+    "just-deploy": "npm run clean && npm run site && bisheng gh-pages --push-only",
+    "deploy": "npm run clean && npm run site && bisheng gh-pages --push-only",
     "pub": "antd-tools run update-self && antd-tools run pub",
     "rn-start": "node node_modules/react-native/local-cli/cli.js start",
     "ios": "concurrently \"antd-tools run watch-tsc\" \"npm run _ios\"",
