+++ site/bisheng.common.config.js
@@ -1,25 +1,7 @@
 const path = require('path');
 const pxtorem = require('postcss-pxtorem');
 
-function pickerGenerator(module) {
-  const tester = new RegExp(`^docs/${module}`);
-  return (markdownData) => {
-    const filename = markdownData.meta.filename;
-    if (tester.test(filename) && !/\.en-US\.md/.test(filename)) {
-      return {
-        meta: markdownData.meta,
-      };
-    }
-  };
-}
-
 module.exports = {
-  port: 8001,
-  source: [
-    './components',
-    './docs',
-    'CHANGELOG.md', // TODO: fix it in bisheng
-  ],
   lazyLoad(nodePath, nodeValue) {
     if (typeof nodeValue === 'string') {
       return true;
@@ -31,39 +13,12 @@ module.exports = {
       const filename = markdownData.meta.filename;
       if (!/^components/.test(filename) ||
           /\/demo$/.test(path.dirname(filename))) return;
+      /* eslint-disable consistent-return */
       return {
         meta: markdownData.meta,
       };
+      /* eslint-enable consistent-return */
     },
-    changelog(markdownData) {
-      if (markdownData.meta.filename === 'CHANGELOG.md') {
-        return {
-          meta: markdownData.meta,
-        };
-      }
-    },
-    'docs/pattern': pickerGenerator('pattern'),
-    'docs/react': pickerGenerator('react'),
-  },
-  entry: {
-    index: {
-      theme: './site/theme',
-      htmlTemplate: './site/theme/static/template.html',
-    },
-    'kitchen-sink': {
-      theme: './site/mobile',
-      htmlTemplate: './site/mobile/static/template.html',
-    },
-  },
-  plugins: [
-    'bisheng-plugin-description',
-    'bisheng-plugin-toc?maxDepth=2',
-    'bisheng-plugin-react?lang=__react',
-    'bisheng-plugin-antd',
-  ],
-  doraConfig: {
-    verbose: true,
-    plugins: ['dora-plugin-upload'],
   },
   webpackConfig(config) {
     config.module.noParse = [/moment.js/];
