+++ site/mobile/static/template.html
@@ -47,7 +47,6 @@
       }
     </script>
     <link rel="icon" href="https://zos.alipayobjects.com/rmsportal/wIjMDnsrDoPPcIV.png" type="image/x-icon">
-    <link rel="stylesheet" type="text/css" href="{{ root }}common.css"/>
     <link rel="stylesheet" type="text/css" href="{{ root }}kitchen-sink.css"/>
   </head>
   <body>
