+++ site/theme/template/Content/ComponentDoc.jsx
@@ -156,8 +156,11 @@ export default class ComponentDoc extends React.Component {
       'code-box-expand-trigger-active': expand,
     });
 
+    const protocol = window.location.protocol;
     const path = doc.meta.filename.split('/')[1];
-    const demoUrl = `${window.location.protocol}//${window.location.host}/kitchen-sink/${path}/`;
+    const isLocalMode = window.location.port;
+    const host = isLocalMode ? 'localhost:8002' : window.location.host;
+    const demoUrl = `${protocol}//${host}/kitchen-sink/${path}/`;
 
     const PopoverContent = (<div>
       <h4 style={{ margin: '8Px 0 12Px' }}>扫二维码查看演示效果</h4>
@@ -166,8 +169,8 @@ export default class ComponentDoc extends React.Component {
 
     const { title, subtitle, chinese, english } = meta;
     const iframeUrl = currentIndex === 0 ?
-    `${window.location.protocol}//${window.location.host}/kitchen-sink/${path}/` :
-    `${window.location.protocol}//${window.location.host}/kitchen-sink/${path}/#${path}-demo-${currentIndex}`;
+            `${protocol}//${host}/${isLocalMode ? '' : 'kitchen-sink'}/${path}/` :
+            `${protocol}//${host}/${isLocalMode ? '' : 'kitchen-sink'}/${path}/#${path}-demo-${currentIndex}`;
     return (
       <DocumentTitle title={`${subtitle || chinese || ''} ${title || english} - Ant Design`}>
         <article>
